auth-service/requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.6.0
Django==5.2.7
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.1
Faker==26.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.1.0
jproperties==2.1.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyfakefs==5.9.3
PyJWT==2.10.1
pysonar==1.2.0.2419
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
responses==0.25.8
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
